Create mathemodica-mps-book.md
---
title: An E-book (Modelica by Application -- Power Systems V1.0)
author: "Atiyah Elsheikh ([Mathemodica.com](http://mathemodica.com/))"
category: "education"
---
The E-book is available as a PDF both for free or for as-much-as-you-think-this-book-deserves. It is available on the collaborative online-latex editor platform overleaf.com. It is hoped that the e-book gets occasionally actualized by interested collaborators.   
 

![The (E-)book Cover](mathemodica-MPSCover1.0.png) 


#### Significant updates since last announcement release 0.3 includes:

* A chapter with the title "System Modeling: A brief History" 
* An initial draft of a section with the title "Commercial Libraries"  
* Sections of topic-related Open Questions to the audience, corresponding to potential extensions to the E-book 

#####  Outline and Ordering details 

More details are on the project website on [github](https://github.com/Mathemodica/ModelicaPowerSystemBook/)

![Mathemodica logo](Mathemodica-logo-50.png)